Update and rename gcd-template.cpp to gcd-snippet.cpp

+++ cp templates/cpp/gcd-snippet.cpp
@@ -1,4 +1,4 @@
-int gcd(int a, int b) {
+long long int gcd(long long int a, long long int b) {
     if (b == 0) return a;
     return gcd(b, a % b);
 }
